my-api/Services/WeatherForecastService.cs
using System;
using my_api.Models;

namespace my_api.Services;

public class WeatherForecastService
{
    private static readonly string[] Summaries = new[]
    {
        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
    };

    public Models.WeatherForecast[] GetWeatherForecasts()
    {
        return Enumerable.Range(1, 5).Select(index =>
            new Models.WeatherForecast
            (
                DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
                Random.Shared.Next(-20, 55),
                Summaries[Random.Shared.Next(Summaries.Length)]
            ))
            .ToArray();
    }
}
